+++ src/kirby/components/grid/grid-card-configuration.ts
@@ -1,0 +1,8 @@
+export class GridCardConfiguration {
+  type: string;
+  preferredSize: number;
+  constructor(type: string, preferredSize: number) {
+    this.type = type;
+    this.preferredSize = preferredSize;
+  }
+}
