+++ src/kirby/components/grid/grid-card.ts
@@ -1,0 +1,15 @@
+import { GridCardConfiguration } from './grid-card-configuration';
+
+
+export class GridCard {
+  configuration: GridCardConfiguration;
+  row: number;
+  col: number;
+  colSpan: number;
+  constructor(configuration: GridCardConfiguration, row: number, col: number, colSpan: number) {
+    this.configuration = configuration;
+    this.row = row;
+    this.col = col;
+    this.colSpan = colSpan;
+  }
+}
